Using size_t instead of unsigned int in MasterBuffer and some other places.  Should help avoid overflow bugs when computing image buffer sizes etc..
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@11461 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ graphics/FilterFunction.cpp
@@ -19,16 +19,16 @@ FilterFunction::~FilterFunction()
 }
 
 
-int FilterFunction::getFilterSpan(int supersample_factor) const
+int FilterFunction::getFilterSpan(size_t supersample_factor) const
 {
-	return (int)ceil(supportRadius() * 2) * supersample_factor;
+	return (int)ceil(supportRadius() * 2) * (int)supersample_factor;
 }
 
 
-void FilterFunction::getFilterDataVec(int supersample_factor, std::vector<float>& filter_data_out) const
+void FilterFunction::getFilterDataVec(size_t supersample_factor, std::vector<float>& filter_data_out) const
 {
 	const double filter_pixel_span_f = supportRadius() * 2 * supersample_factor;
-	const int filter_pixel_span  = (int)ceil(supportRadius() * 2) * supersample_factor;
+	const int filter_pixel_span  = (int)ceil(supportRadius() * 2) * (int)supersample_factor;
 	const int filter_pixel_bound = filter_pixel_span / 2 - 1;
 
 	filter_data_out.resize((filter_pixel_span - 1) * (filter_pixel_span - 1));
